github-backend-main/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiosqlite==0.20.0
pyjwt==2.9.0
bcrypt==4.2.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
beautifulsoup4==4.12.3
playwright==1.47.0
cryptography==42.0.8
aiosmtplib==3.0.2
google-auth==2.23.4
